requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Login==0.6.3
Werkzeug==3.1.0
Flask-WTF==1.2.2
WTForms==3.1.2
python-dotenv==1.0.1
Flask-Limiter==3.8.0
reportlab==4.2.5
psycopg2-binary==2.9.9
